+++ src/components/GenerationPercentageChart/GenerationPercentageChart.tsx
@@ -1,5 +1,7 @@
-
-import React, { useState, useEffect, Fragment, useContext } from 'react'
+import React, { 
+    useState, useEffect, 
+    Fragment, useContext 
+} from 'react'
 
 import config from '../../config';