+++ src/components/shared/Nav.test.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import Nav from './Nav';
+
+describe('<Nav />', () => {
+
+    const renderComponent = () => (<Nav />);
+
+    it('renders without error', async () => {
+        const component = shallow(renderComponent());
+    });
+
+});