+++ CHANGES.md
@@ -13,3 +13,6 @@ https://carbon-intensity.github.io/api-definitions/?javascript--nodejs#get-gener
   - adding .ts to webpack.config.js for transpilation. There was no config for ts(x) files
   - adding .svg extension to file-loader config
 - npm i date-fns
+- npm i enzyme enzyme-adapter-react-16 react-test-renderer enzyme-to-json -D
+- npm i npm install @types/enzyme -D
+-